Move getSiteCommentCounts
