+++ client/my-sites/comments/comment-list/index.jsx
@@ -22,7 +22,7 @@ import QuerySiteCommentCounts from 'components/data/query-site-comment-counts';
 import QuerySiteCommentsList from 'components/data/query-site-comments-list';
 import QuerySiteSettings from 'components/data/query-site-settings';
 import getCommentsPage from 'state/selectors/get-comments-page';
-import getSiteCommentCounts from 'state/selectors/get-site-comment-counts';
+import { getSiteCommentCounts } from 'state/comments/selectors';
 import { bumpStat, composeAnalytics, recordTracksEvent } from 'state/analytics/actions';
 import { COMMENTS_PER_PAGE } from '../constants';
